+++ src/rendering/renderLabelMapImage.js
@@ -6,6 +6,16 @@ import doesImageNeedToBeRendered from './doesImageNeedToBeRendered.js';
 import storedPixelDataToCanvasImageDataColorLUT from '../internal/storedPixelDataToCanvasImageDataColorLUT.js';
 import colors from '../colors/index.js';
 
+/**
+ * Returns an appropriate canvas to render the Image. If the canvas available in the cache is appropriate
+ * it is returned, otherwise adjustments are made. It also sets the color transfer functions.
+ *
+ * @param {Object} enabledElement The cornerstone enabled element
+ * @param {Object} image The image to be rendered
+ * @param {Boolean} invalidated Is pixel data valid
+ * @returns {HTMLCanvasElement} An appropriate canvas for rendering the image
+ * @memberof rendering
+ */
 function getRenderCanvas (enabledElement, image, invalidated) {
   if (!enabledElement.renderingTools.renderCanvas) {
     enabledElement.renderingTools.renderCanvas = document.createElement('canvas');
@@ -13,8 +23,11 @@ function getRenderCanvas (enabledElement, image, invalidated) {
 
   const renderCanvas = enabledElement.renderingTools.renderCanvas;
 
-  // TODO: Deprecate enabledElement.options.colormap
   let colormap = enabledElement.viewport.colormap || enabledElement.options.colormap;
+
+  if (enabledElement.options.colormap) {
+    console.warn('enabledElement.options.colormap is deprecated. Use enabledElement.viewport.colormap instead');
+  }
 
   if (colormap && (typeof colormap === 'string')) {
     colormap = colors.getColormap(colormap);